# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "JOIN API",
    "description": "API мини-приложения **JOIN** в MAX: афиша города, события по Пушкинской карте и поиск компании, чтобы пойти вместе.\n\n### Основной сценарий\n1. **Вход** — `POST /api/auth/login` (или регистрация) возвращает `token`.\n2. **Афиша** — `GET /api/events?city=Казань`, фильтр `pushkinCard=true`.\n3. **Сохранить событие** — `POST /api/events/{id}/like`. Если его уже сохранил подходящий человек, в ответе будет совпадение.\n4. **Позвать пойти вместе** — `POST /api/matches/{id}/request`, второй отвечает `POST /api/matches/{id}/accept`. Только после этого открывается чат.\n5. **Чат** — `GET` / `POST /api/chats/{chatId}/messages`.\n\n### Как попробовать\nПолучите токен через `POST /api/auth/login`, нажмите **Authorize** и вставьте его в `bearerAuth`. Внутри MAX вместо токена приложение передаёт заголовок `X-Max-Init-Data`.\n\n### Ответы\n* Без авторизации защищённые методы отвечают `403` с пустым телом.\n* Ошибки — JSON `ErrorResponse`: `{\"error\": \"текст для пользователя\", \"code\": \"…\"}`.\n* Списки с `page`, `size`, `sort` возвращают страницу: `content`, `totalElements`, `number`, `size`.\n* Даты — `YYYY-MM-DD`, время — `HH:mm:ss`, местное время города события.\n\nОбязательные проверки основного сценария — `DATA-API.yaml` в репозитории.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://2-27-20-14.sslip.io",
      "description": "Демо-стенд"
    },
    {
      "url": "/",
...
    }
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "maxInitData": []
    }
  ],
...
  "paths": {
...
    "/api/users/{id}/block": {
      "get": {
        "tags": [
          "Чёрный список"
        ],
        "summary": "Блокировка между мной и пользователем",
        "description": "`BLOCKED_BY_ME`, `BLOCKED_ME` или ничего.",
        "operationId": "getStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockStatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Чёрный список"
        ],
        "summary": "Заблокировать",
        "description": "Совпадения и переписка с этим человеком прекращаются, в его компанию не вступить.",
        "operationId": "block",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockStatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Чёрный список"
        ],
        "summary": "Разблокировать",
        "operationId": "unblock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockStatusResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/users/me/blocked": {
      "get": {
        "tags": [
          "Чёрный список"
        ],
        "summary": "Мой чёрный список",
        "operationId": "getBlocked",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedUserResponse"
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/universities": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Вузы",
        "description": "Справочник вузов России; `city` — только вузы города.",
        "operationId": "getUniversities",
        "parameters": [
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UniversityResponse"
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/universities/{id}": {
      "get": {
        "tags": [
          "Справочники"
        ],
        "summary": "Вуз",
        "operationId": "getUniversity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniversityResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Что пошло не так",
            "example": "Сегодня можно лайкнуть не больше 10 событий — возвращайся завтра"
          },
          "message": {
            "type": "string",
            "description": "То же, что `error` (есть у ошибок действий пользователя, код 409)"
          },
          "code": {
            "type": "string",
            "description": "Машинный код ошибки, если есть",
            "example": "VALIDATION_FAILED"
          }
        },
        "description": "Ошибка. Текст в `error` можно показывать пользователю"
      },
...
      "BlockStatusResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "blockedByMe": {
            "type": "boolean"
          },
          "blockedMe": {
            "type": "boolean"
          }
        }
      },
...
      "BlockedUserResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "blockedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UniversityResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          }
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "Нет авторизации (тело пустое) или нет доступа к объекту"
      },
      "NotFound": {
        "description": "Объект не найден или недоступен",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Event not found with id: 42"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Токен из POST /api/auth/login или /api/auth/register",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "maxInitData": {
        "type": "apiKey",
        "description": "window.WebApp.initData из MAX Bridge; подпись проверяется токеном бота",
        "name": "X-Max-Init-Data",
        "in": "header"
      }
    }
  }
}